+++ src/mpv/70_services.js
@@ -665,6 +665,7 @@ IINATAN.updateSubtitleGeometry = function () {
         });
         IINATAN.state.subtitleUnits = combined;
         IINATAN.state.subtitleRect = IINATAN.unionRects(rects);
+        IINATAN.handleHover();
       },
       10000,
     );